Discounting factor for reward
The sum of the rewards is now calculated with a discounting factor. This means that immediate rewards have a higher weight. For eacht timestep in X there is also now a new column which contains the discounted reward.
This will later be used to filter out good training instances by selecting only those with a score higher than a minimum value.